Fixed an issue with loading files
Previously, when loading a file of length N, and letting it
run to its exit-statement would, when loading a file of a length
less than N trigger an out-of-bounds exception on the table
model in the "ProgramView" class. That problem has been fixed.

The cause was not resetting the currentRowIndex variable